requirements.txt
openai==1.25.0
langchain==0.1.17
langchain-openai==0.1.5
faiss-cpu==1.8.0
tiktoken == 0.6.0
streamlit_chat==0.1.1
streamlit==1.33.0
PyPDF2== 3.0.1
python-docx==0.8.11
